# Remove unsafe in with_nix_path() for [u8]